moved excepthook to cs50.py
