+++ src/cs50/cs50.py
@@ -1,7 +1,33 @@
 from __future__ import print_function
+
 import inspect
 import re
 import sys
+
+from os.path import abspath, join
+from site import getsitepackages, getusersitepackages
+from termcolor import cprint
+from traceback import extract_tb, format_list, format_exception_only
+
+
+def excepthook(type, value, tb):
+    """
+    Format traceback, darkening entries from global site-packages and user-specific site-packages directory.
+
+    https://stackoverflow.com/a/33042323/5156190
+    """
+    packages = tuple(join(abspath(p), "") for p in getsitepackages() + [getusersitepackages()])
+    for entry in extract_tb(tb):
+        fmt = format_list((entry,))
+        if (entry.filename.startswith(packages)):
+            cprint("".join(fmt), attrs=["dark"], end="", file=sys.stderr)
+        else:
+            cprint("".join(fmt), end="", file=sys.stderr)
+    cprint("".join(format_exception_only(type, value)), end="")
+
+
+sys.excepthook = excepthook
+
 
 class flushfile():
     """
@@ -9,6 +35,7 @@ class flushfile():
 
     http://stackoverflow.com/a/231216
     """
+
     def __init__(self, f):
         self.f = f
 
@@ -18,8 +45,11 @@ class flushfile():
     def write(self, x):
         self.f.write(x)
         self.f.flush()
+
+
 sys.stderr = flushfile(sys.stderr)
 sys.stdout = flushfile(sys.stdout)
+
 
 def eprint(*args, **kwargs):
     """
@@ -31,6 +61,7 @@ def eprint(*args, **kwargs):
     (filename, lineno) = inspect.stack()[1][1:3]
     print("{}:{}: ".format(filename, lineno), end="")
     print(*args, end=end, file=sys.stderr, sep=sep)
+
 
 def get_char(prompt=None):
     """
@@ -48,6 +79,7 @@ def get_char(prompt=None):
         # temporarily here for backwards compatibility
         if prompt is None:
             print("Retry: ", end="")
+
 
 def get_float(prompt=None):
     """
@@ -69,6 +101,7 @@ def get_float(prompt=None):
         if prompt is None:
             print("Retry: ", end="")
 
+
 def get_int(prompt=None):
     """
     Read a line of text from standard input and return the equivalent int;
@@ -76,13 +109,13 @@ def get_int(prompt=None):
     can't be read, return None.
     """
     while True:
-        s = get_string(prompt);
+        s = get_string(prompt)
         if s is None:
             return None
         if re.search(r"^[+-]?\d+$", s):
             try:
                 i = int(s, 10)
-                if type(i) is int: # could become long in Python 2
+                if type(i) is int:  # could become long in Python 2
                     return i
             except ValueError:
                 pass
@@ -90,6 +123,7 @@ def get_int(prompt=None):
         # temporarily here for backwards compatibility
         if prompt is None:
             print("Retry: ", end="")
+
 
 if sys.version_info.major != 3:
     def get_long(prompt=None):
@@ -111,6 +145,7 @@ if sys.version_info.major != 3:
             # temporarily here for backwards compatibility
             if prompt is None:
                 print("Retry: ", end="")
+
 
 def get_string(prompt=None):
     """
